+++ frontend/src/App.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
 import { createStackNavigator } from '@react-navigation/stack';
-import HomeScreen from './home-screen';
-import TravelDiaryScreen from './travel-diary-screen';
+import HomeScreen from './screens/home-screen';
+import TravelDiaryScreen from './screens/travel-diary-screen';
 
 const Stack = createStackNavigator();
