+++ frontend/src/assets/diary-icon.png
