+++ frontend/src/assets/splash-icon.png
